requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-crispy-forms==1.7.2
django-filter==2.0.0
django==2.1
djangorestframework-jsonapi==2.5.0
djangorestframework==3.8.2
inflection==0.3.1         # via djangorestframework-jsonapi
markdown==2.6.11
pytz==2018.5              # via django
six==1.11.0               # via djangorestframework-jsonapi
